khoa commit

+++ resources/views/company/list.blade.php
@@ -46,7 +46,7 @@
                     <td>
                         <a class="btn btn-info" href="/system-admin/company/detail/{{$company->id}}" role="button">Detail</a>
                         <a class="btn btn-primary" href="/system-admin/company/update/{{$company->id}}" role="button">Update</a>
-                        <a class="btn btn-danger" href="/system-admin/company/delete/{{$company->id}}" role="button">Delete</a>
+                        <a onclick="return confirm('Are you sure you want to delete this company?');" class="btn btn-danger" href="/system-admin/company/delete/{{$company->id}}" role="button">Delete</a>
                     </td>
                 </tr>
             @endforeach
